integration-tests: big refactor and stabilise local tests
