+++ integration-tests/cli/jobs/simple.js
@@ -1,1 +1,2 @@
+console.log(' *** wibble *** ');
 fn(() => 42);
