+++ integration-tests/cli/src/run.ts
@@ -3,31 +3,16 @@ import * as path from 'node:path';
 
 const isProd = process.env.NODE_ENV === 'production';
 
-console.log();
-console.log('== openfn CLI integration tests ==\n');
-if (isProd) {
-  console.log('Running tests in Production mode');
-  console.log('Tests will use the global openfn command');
-} else {
-  console.log('Running tests in dev mode');
-  console.log('Tests will use the local build in kit/packages/cli');
-}
-console.log();
-
-const run = async (args) => {
-  const exe = isProd
-    ? // In production, use the global
-      'openfn'
-    : // But in dev, use this repo's CLI
-      'pnpm -C packages/cli openfn';
-
-  const cmd = `${exe} ${args}`;
+const run = async (cmd: string) => {
+  if (!isProd) {
+    cmd = cmd.replace(/^openfn/, 'pnpm -C ../../packages/cli openfn');
+  }
   return new Promise((resolve) => {
     const options = {
       // cwd: isProd ? '.' : path.resolve('../..'),
       env: {
         ...process.env,
-        OPENFN_REPO_HOME: 'repo',
+        OPENFN_REPO_DIR: path.resolve('repo'),
       },
     };
     exec(cmd, options, (err, stdout, stderr) => {
