+++ integration-tests/cli/test/cli.test.ts
@@ -1,10 +1,11 @@
 import test from 'ava';
-import run from './run';
+import run from '../src/run';
 
 test.serial('version', async (t) => {
   const { stdout } = await run('openfn version');
+
   t.regex(stdout, /Versions/);
-  t.regex(stdout, /Node.js/);
+  t.regex(stdout, /node.js/);
   t.regex(stdout, /cli/);
   t.regex(stdout, /runtime/);
   t.regex(stdout, /compiler/);
