--- packages/integration-tests/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "ts-node": {
-    "experimentalSpecifierResolution": "node"
-  },
-  "compilerOptions": {
-    "module": "es2020",
-    "moduleResolution": "node",
-    "allowJs": true,
-    "isolatedModules": true,
-    "noEmit": true,
-    "skipLibCheck": true
-  }
-}
