+++ pnpm-workspace.yaml
@@ -2,5 +2,6 @@ packages:
   # all packages in subdirs of packages/ and components/
   - 'packages/**'
   - 'examples/**'
+  - 'integration-tests/**'
   # exclude packages that are inside test directories
   - '!**/test/**'
